requirements.txt
diff-match-patch==20121119
Django==1.9.4
django-filter==0.12.0
django-import-export==0.4.3
djangorestframework==3.3.2
tablib==0.11.2
mysqlclient==1.3.7
beautifulsoup4
pytz
django-ajax-selects

